League of Legends champion counter
Was talking to some friends in my summer vacation and i had a friend mention how he would love a small app that tells him who to pick against another champion in the Riot game League of Legends. So i decided to create a fun little Tkinter Python project. Haven't made it an .exe file yet. But soon to come! . hope you enjoy it. Notice you need downloaded libraries if you are running IDLE instead of PyCharm which is the IDE i use for developing. Enjoy!